remove activateAccount controller to user.js
