+++ server/src/controllers/index.js
@@ -1,11 +1,14 @@
 const { clientError, serverError } = require('./error');
 const { signup, login, logout } = require('./auth');
-// const login = require('./login');
 const { getRBookingbyDate, bookingRoom, deleteBooking } = require('./booking');
-// const logout = require('./logout');
 const { addRoom, getRooms, patchRoom, deleteRoomById } = require('./room');
-const { deleteUser, getUsers, getProfile, patchProfile } = require('./users');
-const activateAccount = require('./activateAccount');
+const {
+  deleteUser,
+  getUsers,
+  getProfile,
+  patchProfile,
+  activateAccount,
+} = require('./users');
 
 module.exports = {
   clientError,
