+++ server/src/controllers/validation/activateSchema.js
@@ -1,0 +1,6 @@
+const Joi = require('@hapi/joi');
+
+module.exports = Joi.object({
+  active: Joi.boolean(),
+  admin: Joi.boolean(),
+});
